Initial Commit - Setup basic project structure, GLFW, GLM, OpenGL and CMake Configuration
Had to modify the CMakeList.txt so that CMake could find the GLFW and GLM libraries.  I didn't want to use a package handler as it felt like better experience to manage it myself.  Only issue is it took WAY too long to get everything sorted, built and in a logical place so that I could replicate the files easily on other systems if I were working on multiple machines.  I will probably figure out how to use vcpkg next time.